requirements.txt
dj-database-url==0.5.0
Django==2.1.1
django-heroku==0.3.1
django-shortuuidfield==0.1.3
psycopg2==2.7.5
pytz==2018.5
shortuuid==0.5.0
six==1.11.0
whitenoise==4.0
